# hermes/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hermes Autonomous AI Assistant API
  description: |
    Comprehensive API documentation for the Hermes Autonomous AI Assistant.

    Hermes is an advanced AI system that combines meta-learning, backend-agnostic architecture,
    and autonomous operations to provide intelligent assistance with self-improving capabilities.

    ## Features
    - **Meta-Learning Engine**: Intelligent question generation and pattern recognition
    - **Backend Management**: Support for multiple AI providers with intelligent routing
    - **Autonomous Operations**: Self-scheduling, suggestion engine, and context automation
    - **Self-Improving System**: Code modification, autonomous learning, and self-healing
    - **Advanced Analytics**: AI-powered insights and evidence-based recommendations

    ## Authentication
    The API uses Bearer token authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer your-api-token-here
    ```
  version: 1.0.0
  contact:
    name: Hermes AI Team
    email: api@hermes-ai.com
    url: https://hermes-ai.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.hermes-ai.com/v1
    description: Production server
  - url: https://staging-api.hermes-ai.com/v1
    description: Staging server
  - url: http://localhost:8080/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /orchestrator/execute:
    post:
      tags:
        - Orchestrator
      summary: Execute intelligent task
      description: |
        Executes an intelligent task using the Hermes AI Assistant with full meta-learning
        and autonomous operation capabilities.
      operationId: executeTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
            examples:
              simple_task:
                summary: Simple analysis task
                value:
                  conversation_id: "conv_1234567890"
                  idea: "Analyze user behavior patterns"
                  context:
                    domain: "analytics"
                    priority: "normal"
              complex_task:
                summary: Complex development task
                value:
                  conversation_id: "conv_1234567890"
                  idea: "Optimize database queries for better performance"
                  context:
                    domain: "software_development"
                    priority: "high"
                    constraints:
                      max_execution_time: 300
                      budget_cents: 500
                    requirements:
                      include_code_examples: true
                      detailed_explanation: true
      responses:
        '200':
          description: Task executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
              examples:
                success:
                  summary: Successful task execution
                  value:
                    success: true
                    data:
                      task_id: "task_1234567890"
                      conversation_id: "conv_1234567890"
                      result:
                        summary: "Analysis completed successfully"
                        details: "User behavior patterns identified and documented"
                        suggestions: ["Implement A/B testing", "Add user feedback collection"]
                        confidence_score: 0.87
                      execution_metadata:
                        backend_used: "openai_gpt4"
                        execution_time_ms: 2340
                        cost_cents: 15
                        tokens_used: 1250
                    message: "Task executed successfully"
                    timestamp: "2024-01-01T12:05:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orchestrator/status:
    get:
      tags:
        - Orchestrator
      summary: Get system status
      description: |
        Returns the current status of the Hermes system including health checks,
        active backends, and system metrics.
      operationId: getSystemStatus
      parameters:
        - name: include_backends
          in: query
          description: Include detailed backend status
          required: false
          schema:
            type: boolean
            default: false
        - name: include_metrics
          in: query
          description: Include system metrics
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: System status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatusResponse'
              examples:
                basic:
                  summary: Basic system status
                  value:
                    success: true
                    data:
                      status: "healthy"
                      version: "1.0.0"
                      uptime_seconds: 86400
                      active_conversations: 45
                    message: "System status retrieved successfully"
                    timestamp: "2024-01-01T12:00:00Z"
                detailed:
                  summary: Detailed system status with backends
                  value:
                    success: true
                    data:
                      status: "healthy"
                      version: "1.0.0"
                      uptime_seconds: 86400
                      active_conversations: 45
                      backends:
                        openai_gpt4:
                          status: "healthy"
                          response_time_ms: 1200
                          success_rate: 0.98
                          cost_per_1k_tokens: 0.03
                        anthropic_claude:
                          status: "healthy"
                          response_time_ms: 1500
                          success_rate: 0.97
                          cost_per_1k_tokens: 0.015
                      metrics:
                        cpu_usage_percent: 45.2
                        memory_usage_percent: 67.8
                        requests_per_minute: 120
                        average_response_time_ms: 1350
                    message: "System status retrieved successfully"
                    timestamp: "2024-01-01T12:00:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExecuteRequest:
      type: object
      required:
        - conversation_id
        - idea
      properties:
        conversation_id:
          type: string
          description: Conversation ID from previous conversation creation
          example: "conv_1234567890"
        idea:
          type: string
          description: The task or idea to execute
          example: "Analyze user behavior patterns and suggest improvements"
        context:
          $ref: '#/components/schemas/ExecutionContext'
    ExecutionContext:
      type: object
      properties:
        domain:
          type: string
          description: Domain of the task
          example: "analytics"
        priority:
          type: string
          enum: [low, normal, high, urgent]
          description: Task priority
          example: "high"
        constraints:
          $ref: '#/components/schemas/TaskConstraints'
        requirements:
          $ref: '#/components/schemas/TaskRequirements'
    TaskConstraints:
      type: object
      properties:
        max_execution_time:
          type: integer
          description: Maximum execution time in seconds
          example: 300
        budget_cents:
          type: integer
          description: Maximum budget in cents
          example: 500
        backends:
          type: array
          items:
            type: string
          description: Preferred backends to use
          example: ["openai_gpt4", "anthropic_claude"]
    TaskRequirements:
      type: object
      properties:
        include_code_examples:
          type: boolean
          description: Include code examples in response
          example: true
        detailed_explanation:
          type: boolean
          description: Provide detailed explanations
          example: true
        references:
          type: array
          items:
            type: string
          description: Include references or sources
          example: ["official_docs", "best_practices"]
    ExecuteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            task_id:
              type: string
              example: "task_1234567890"
            conversation_id:
              type: string
              example: "conv_1234567890"
            result:
              $ref: '#/components/schemas/TaskResult'
            execution_metadata:
              $ref: '#/components/schemas/ExecutionMetadata'
        message:
          type: string
          example: "Task executed successfully"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:05:00Z"
    TaskResult:
      type: object
      properties:
        summary:
          type: string
          description: Brief summary of the result
          example: "Analysis completed successfully"
        details:
          type: string
          description: Detailed result explanation
          example: "User behavior patterns identified and documented"
        suggestions:
          type: array
          items:
            type: string
          description: Generated suggestions
          example: ["Implement A/B testing", "Add user feedback collection"]
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence in the result
          example: 0.87
        next_steps:
          type: array
          items:
            type: string
          description: Recommended next steps
          example: ["Implement tracking", "Monitor metrics"]
    ExecutionMetadata:
      type: object
      properties:
        backend_used:
          type: string
          description: Backend used for execution
          example: "openai_gpt4"
        execution_time_ms:
          type: integer
          description: Execution time in milliseconds
          example: 2340
        cost_cents:
          type: integer
          description: Cost in cents
          example: 15
        tokens_used:
          type: integer
          description: Number of tokens used
          example: 1250
        optimization_applied:
          type: boolean
          description: Whether optimization was applied
          example: true
    SystemStatusResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            status:
              type: string
              enum: [healthy, degraded, unhealthy]
              example: "healthy"
            version:
              type: string
              example: "1.0.0"
            uptime_seconds:
              type: integer
              example: 86400
            active_conversations:
              type: integer
              example: 45
            backends:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/BackendStatus'
            metrics:
              $ref: '#/components/schemas/SystemMetrics'
        message:
          type: string
          example: "System status retrieved successfully"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
    BackendStatus:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy, unavailable]
          example: "healthy"
        response_time_ms:
          type: integer
          example: 1200
        success_rate:
          type: number
          format: float
          example: 0.98
        cost_per_1k_tokens:
          type: number
          format: float
          example: 0.03
        last_check:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
    SystemMetrics:
      type: object
      properties:
        cpu_usage_percent:
          type: number
          format: float
          example: 45.2
        memory_usage_percent:
          type: number
          format: float
          example: 67.8
        requests_per_minute:
          type: integer
          example: 120
        average_response_time_ms:
          type: integer
          example: 1350
        error_rate:
          type: number
          format: float
          example: 0.02
...
  responses:
    BadRequest:
      description: Bad request - Invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Invalid request parameters"
              details:
                type: object
                example:
                  field: "conversation_id"
                  message: "Required field missing"
              timestamp:
                type: string
                format: date-time
                example: "2024-01-01T12:00:00Z"
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Unauthorized access"
              message:
                type: string
                example: "Valid authentication token required"
              timestamp:
                type: string
                format: date-time
                example: "2024-01-01T12:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Resource not found"
              message:
                type: string
                example: "The requested resource was not found"
              timestamp:
                type: string
                format: date-time
                example: "2024-01-01T12:00:00Z"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Rate limit exceeded"
              message:
                type: string
                example: "Too many requests. Please try again later."
              retry_after:
                type: integer
                example: 60
              timestamp:
                type: string
                format: date-time
                example: "2024-01-01T12:00:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Internal server error"
              message:
                type: string
                example: "An unexpected error occurred"
              request_id:
                type: string
                example: "req_1234567890"
              timestamp:
                type: string
                format: date-time
                example: "2024-01-01T12:00:00Z"
...